Rescope and move classes in no-player where necessary.
