+++ core/src/main/java/com/novoda/noplayer/NoPlayerView.java
@@ -8,7 +8,6 @@ import android.widget.FrameLayout;
 
 import com.google.android.exoplayer2.ui.AspectRatioFrameLayout;
 import com.google.android.exoplayer2.ui.SubtitleView;
-import com.novoda.noplayer.exoplayer.AspectRatioChangeCalculator;
 
 public class NoPlayerView extends FrameLayout implements AspectRatioChangeCalculator.Listener, PlayerView {
 
@@ -84,7 +83,7 @@ public class NoPlayerView extends FrameLayout implements AspectRatioChangeCalcul
     private final Player.VideoSizeChangedListener videoSizeChangedListener = new Player.VideoSizeChangedListener() {
         @Override
         public void onVideoSizeChanged(int width, int height, int unappliedRotationDegrees, float pixelWidthHeightRatio) {
-            aspectRatioChangeCalculator.onVideoSizeChanged(width, height, unappliedRotationDegrees, pixelWidthHeightRatio);
+            aspectRatioChangeCalculator.onVideoSizeChanged(width, height, pixelWidthHeightRatio);
         }
     };
