+++ core/src/main/java/com/novoda/noplayer/exoplayer/forwarder/PlayerOnErrorForwarder.java
@@ -7,7 +7,7 @@ import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.novoda.noplayer.Player;
-import com.novoda.noplayer.exoplayer.ExoPlayerErrorFactory;
+import com.novoda.noplayer.exoplayer.playererror.ExoPlayerErrorFactory;
 import com.novoda.noplayer.listeners.ErrorListeners;
 
 class PlayerOnErrorForwarder implements ExoPlayer.EventListener {
